requirements.txt
# Core Django
Django==5.2.9
django-environ==0.12.1
dj-database-url==3.1.0
psycopg2-binary==2.9.11
gunicorn==23.0.0
whitenoise==6.11.0

# Telegram Bot
python-telegram-bot==20.7
python-dotenv==1.2.1

# Web Scraping
playwright==1.58.0
beautifulsoup4==4.14.2
aiohttp==3.13.3

# AI/ML (Gemini only)
google-generativeai
google-ai-generativelanguage

# PDF Generation
fpdf2==2.8.5

# Utilities
loguru==0.7.2
requests==2.32.5
python-dateutil==2.9.0.post0
pytz==2025.2
